requirements.txt
django
djangorestframework
django-filter
mock
Pillow
psycopg2